Finalize split up
