+++ internal/extsvc/github/common_test.go
@@ -1,0 +1,15 @@
+package github
+
+import (
+	"flag"
+	"regexp"
+)
+
+var updateRegex = flag.String("update", "", "Update testdata of tests matching the given regex")
+
+func update(name string) bool {
+	if updateRegex == nil || *updateRegex == "" {
+		return false
+	}
+	return regexp.MustCompile(*updateRegex).MatchString(name)
+}
