+++ internal/extsvc/github/user_test.go
@@ -9,7 +9,7 @@ import (
 
 // NOTE: To update VCR for this test, please use the token of "sourcegraph-vcr"
 // for GITHUB_TOKEN, which can be found in 1Password.
-func TestClient_ListRepositoryCollaborators(t *testing.T) {
+func TestListRepositoryCollaborators(t *testing.T) {
 	tests := []struct {
 		name      string
 		owner     string
@@ -50,7 +50,7 @@ func TestClient_ListRepositoryCollaborators(t *testing.T) {
 	}
 	for _, test := range tests {
 		t.Run(test.name, func(t *testing.T) {
-			client, save := newClient(t, "ListRepositoryCollaborators_"+test.name)
+			client, save := newV3Client(t, "ListRepositoryCollaborators_"+test.name)
 			defer save()
 
 			users, _, err := client.ListRepositoryCollaborators(context.Background(), test.owner, test.repo, 1)
