+++ internal/extsvc/github/v3_test.go
@@ -1,0 +1,206 @@
+package github
+
+import (
+	"context"
+	"net/url"
+	"os"
+	"testing"
+
+	"github.com/google/go-cmp/cmp"
+	"github.com/sourcegraph/sourcegraph/internal/extsvc/auth"
+	"github.com/sourcegraph/sourcegraph/internal/httptestutil"
+	"github.com/sourcegraph/sourcegraph/internal/rcache"
+	"github.com/sourcegraph/sourcegraph/internal/testutil"
+)
+
+func TestNewRepoCache(t *testing.T) {
+	cmpOpts := cmp.AllowUnexported(rcache.Cache{})
+	t.Run("GitHub.com", func(t *testing.T) {
+		url, _ := url.Parse("https://www.github.com")
+		token := &auth.OAuthBearerToken{Token: "asdf"}
+
+		// github.com caches should:
+		// (1) use githubProxyURL for the prefix hash rather than the given url
+		// (2) have a TTL of 10 minutes
+		prefix := "gh_repo:" + token.Hash()
+		got := newRepoCache(url, token)
+		want := rcache.NewWithTTL(prefix, 600)
+		if diff := cmp.Diff(want, got, cmpOpts); diff != "" {
+			t.Fatal(diff)
+		}
+	})
+
+	t.Run("GitHub Enterprise", func(t *testing.T) {
+		url, _ := url.Parse("https://www.sourcegraph.com")
+		token := &auth.OAuthBearerToken{Token: "asdf"}
+
+		// GitHub Enterprise caches should:
+		// (1) use the given URL for the prefix hash
+		// (2) have a TTL of 30 seconds
+		prefix := "gh_repo:" + token.Hash()
+		got := newRepoCache(url, token)
+		want := rcache.NewWithTTL(prefix, 30)
+		if diff := cmp.Diff(want, got, cmpOpts); diff != "" {
+			t.Fatal(diff)
+		}
+	})
+}
+
+// NOTE: To update VCR for this test, please use the token of "sourcegraph-vcr"
+// for GITHUB_TOKEN, which can be found in 1Password.
+func TestListAffiliatedRepositories(t *testing.T) {
+	tests := []struct {
+		name       string
+		visibility Visibility
+		wantRepos  []*Repository
+	}{
+		{
+			name:       "list all repositories",
+			visibility: VisibilityAll,
+			wantRepos: []*Repository{
+				{
+					ID:               "MDEwOlJlcG9zaXRvcnkyNjMwMzQxNTE=",
+					DatabaseID:       263034151,
+					NameWithOwner:    "sourcegraph-vcr-repos/private-org-repo-1",
+					URL:              "https://github.com/sourcegraph-vcr-repos/private-org-repo-1",
+					IsPrivate:        true,
+					ViewerPermission: "ADMIN",
+				}, {
+					ID:               "MDEwOlJlcG9zaXRvcnkyNjMwMzQwNzM=",
+					DatabaseID:       263034073,
+					NameWithOwner:    "sourcegraph-vcr/private-user-repo-1",
+					URL:              "https://github.com/sourcegraph-vcr/private-user-repo-1",
+					IsPrivate:        true,
+					ViewerPermission: "ADMIN",
+				}, {
+					ID:               "MDEwOlJlcG9zaXRvcnkyNjMwMzM5NDk=",
+					DatabaseID:       263033949,
+					NameWithOwner:    "sourcegraph-vcr/public-user-repo-1",
+					URL:              "https://github.com/sourcegraph-vcr/public-user-repo-1",
+					ViewerPermission: "ADMIN",
+				}, {
+					ID:               "MDEwOlJlcG9zaXRvcnkyNjMwMzM3NjE=",
+					DatabaseID:       263033761,
+					NameWithOwner:    "sourcegraph-vcr-repos/public-org-repo-1",
+					URL:              "https://github.com/sourcegraph-vcr-repos/public-org-repo-1",
+					ViewerPermission: "ADMIN",
+				},
+			},
+		},
+		{
+			name:       "list public repositories",
+			visibility: VisibilityPublic,
+			wantRepos: []*Repository{
+				{
+					ID:               "MDEwOlJlcG9zaXRvcnkyNjMwMzM5NDk=",
+					DatabaseID:       263033949,
+					NameWithOwner:    "sourcegraph-vcr/public-user-repo-1",
+					URL:              "https://github.com/sourcegraph-vcr/public-user-repo-1",
+					ViewerPermission: "ADMIN",
+				}, {
+					ID:               "MDEwOlJlcG9zaXRvcnkyNjMwMzM3NjE=",
+					DatabaseID:       263033761,
+					NameWithOwner:    "sourcegraph-vcr-repos/public-org-repo-1",
+					URL:              "https://github.com/sourcegraph-vcr-repos/public-org-repo-1",
+					ViewerPermission: "ADMIN",
+				},
+			},
+		},
+		{
+			name:       "list private repositories",
+			visibility: VisibilityPrivate,
+			wantRepos: []*Repository{
+				{
+					ID:               "MDEwOlJlcG9zaXRvcnkyNjMwMzQxNTE=",
+					DatabaseID:       263034151,
+					NameWithOwner:    "sourcegraph-vcr-repos/private-org-repo-1",
+					URL:              "https://github.com/sourcegraph-vcr-repos/private-org-repo-1",
+					IsPrivate:        true,
+					ViewerPermission: "ADMIN",
+				}, {
+					ID:               "MDEwOlJlcG9zaXRvcnkyNjMwMzQwNzM=",
+					DatabaseID:       263034073,
+					NameWithOwner:    "sourcegraph-vcr/private-user-repo-1",
+					URL:              "https://github.com/sourcegraph-vcr/private-user-repo-1",
+					IsPrivate:        true,
+					ViewerPermission: "ADMIN",
+				},
+			},
+		},
+	}
+	for _, test := range tests {
+		t.Run(test.name, func(t *testing.T) {
+			client, save := newV3Client(t, "ListAffiliatedRepositories_"+test.name)
+			defer save()
+
+			repos, _, _, err := client.ListAffiliatedRepositories(context.Background(), test.visibility, 1)
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			if diff := cmp.Diff(test.wantRepos, repos); diff != "" {
+				t.Fatalf("Repos mismatch (-want +got):\n%s", diff)
+			}
+		})
+	}
+}
+
+func TestGetAuthenticatedUserOrgs(t *testing.T) {
+	cli, save := newV3Client(t, "GetAuthenticatedUserOrgs")
+	defer save()
+
+	ctx := context.Background()
+	orgs, err := cli.GetAuthenticatedUserOrgs(ctx)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	testutil.AssertGolden(t,
+		"testdata/golden/GetAuthenticatedUserOrgs",
+		update("GetAuthenticatedUserOrgs"),
+		orgs,
+	)
+}
+
+func TestV3Client_WithAuthenticator(t *testing.T) {
+	uri, err := url.Parse("https://github.com")
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	old := &V3Client{
+		apiURL: uri,
+		auth:   &auth.OAuthBearerToken{Token: "old_token"},
+	}
+
+	newToken := &auth.OAuthBearerToken{Token: "new_token"}
+	new := old.WithAuthenticator(newToken)
+	if old == new {
+		t.Fatal("both clients have the same address")
+	}
+
+	if new.auth != newToken {
+		t.Fatalf("token: want %q but got %q", newToken, new.auth)
+	}
+}
+
+func newV3Client(t testing.TB, name string) (*V3Client, func()) {
+	t.Helper()
+
+	cf, save := httptestutil.NewGitHubRecorderFactory(t, update(name), name)
+	uri, err := url.Parse("https://github.com")
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	doer, err := cf.Doer()
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	cli := NewV3Client(uri, &auth.OAuthBearerToken{
+		Token: os.Getenv("GITHUB_TOKEN"),
+	}, doer)
+
+	return cli, save
+}
